requirements.txt
mpmath>=1.3.0
numpy>=1.26.4
scikit-learn>=1.5.1
tqdm>=4.66.5
torchvision>=0.19.1
matplotlib>=3.9.2
scipy>=1.14.1